feat/ refactor naming and packages
