+++ ldes-server-domain/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/domain/ldesconfig/config/LdesAdminConstants.java
@@ -1,11 +1,13 @@
-package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.config;
+package be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.config;
 
 public class LdesAdminConstants {
 	public static final String VIEW_NAME = "viewName";
 	public static final String TREE = "https://w3id.org/tree#";
 	public static final String LDES = "https://w3id.org/ldes#";
-	public static final String VIEW = LDES + "View";
-	public static final String SHAPE = TREE + "Shape";
+	public static final String VIEW = LDES + "view";
+	public static final String VIEW_TYPE = LDES + "View";
+	public static final String SHAPE = TREE + "shape";
+	public static final String NODE_SHAPE_TYPE = TREE + "NodeShape";
 	public static final String EVENT_STREAM_TYPE = LDES + "EventStream";
 	public static final String NAME_PREFIX = "prefix";
