+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-1.ttl
@@ -3,4 +3,4 @@
 
 ldes:collectionName1 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape [] .
+   tree:shape [] .
