+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-3.ttl
@@ -3,8 +3,8 @@
 
 ldes:collectionName3 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape [] ;
-   ldes:View [
+   tree:shape [] ;
+   ldes:view [
        tree:viewDescription [
            ldes:retentionPolicy "none"
        ]
