+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-multiple-views.ttl
@@ -3,18 +3,18 @@
 
 ldes:collectionName2 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape [] ;
-   ldes:View [
+   tree:shape [] ;
+   ldes:view [
        tree:viewDescription [
            ldes:retentionPolicy "none"
        ]
    ] ;
-   ldes:View [
+   ldes:view [
        tree:viewDescription [
            ldes:retentionPolicy "none"
        ]
    ] ;
-   ldes:View [
+   ldes:view [
        tree:viewDescription [
            ldes:retentionPolicy "none"
        ]
