+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-named-view.ttl
@@ -3,8 +3,8 @@
 
 ldes:collectionName1 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape [] ;
-   ldes:View ldes:view1 .
+   tree:shape [] ;
+   ldes:view ldes:view1 .
 
 ldes:view1 a ldes:View ;
    tree:viewDescription [
