+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-with-named-view.ttl
@@ -4,10 +4,10 @@
 
 ldes:collectionName1 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape ldes:shapeNew ;
-   ldes:View ldes:view1 .
+   tree:shape ldes:shapeNew ;
+   ldes:view ldes:view1 .
 
-ldes:shapeNew a sh:nodeShape ;
+ldes:shapeNew a sh:NodeShape ;
    sh:nodeShape [
      sh:closed "true" ;
      sh:propertyShape []
