+++ ldes-server-domain/src/test/resources/eventstream/streams/ldes-with-shape.ttl
@@ -4,7 +4,7 @@
 
 ldes:collectionName1 a ldes:EventStream ;
    ldes:versionOfPath "1.1" ;
-   ldes:Shape [ a sh:nodeShape ;
+   tree:shape [ a sh:nodeShape ;
       sh:nodeShape [
          sh:closed "true" ;
          ] ;
@@ -13,4 +13,4 @@ ldes:collectionName1 a ldes:EventStream ;
          ] ;
       sh:deactivated "false" ;
    ] ;
-   ldes:View ldes:view1 .
+   ldes:view ldes:view1 .
