+++ ldes-server-port-admin-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/config/AdminWebConfig.java
@@ -7,14 +7,15 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class AdminWebConfig {
-    @Bean
-    public LdesConfigModelConverter ldesStreamModelConverter() {
-        return new LdesConfigModelConverter();
-    }
-    @Bean("streamShaclValidator")
-    public LdesConfigShaclValidator ldesStreamShaclValidator() {
-        return new LdesConfigShaclValidator("streamShaclShape.ttl");
-    }
+	@Bean
+	public LdesConfigModelConverter ldesStreamModelConverter() {
+		return new LdesConfigModelConverter();
+	}
+
+	@Bean("streamShaclValidator")
+	public LdesConfigShaclValidator ldesStreamShaclValidator() {
+		return new LdesConfigShaclValidator("streamShaclShape.ttl");
+	}
 
 	@Bean(name = "viewShaclValidator")
 	public LdesConfigShaclValidator ldesViewShaclValidator() {
