+++ ldes-server-port-admin-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/controllers/AdminShapeRestController.java
@@ -1,9 +1,8 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.controllers;
 
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.services.LdesConfigModelService;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.valueobjects.LdesConfigModel;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.services.LdesConfigModelService;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.valueobjects.LdesConfigModel;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.validation.LdesConfigShaclValidator;
-import org.apache.jena.rdf.model.Model;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/admin/api/v1")
 public class AdminShapeRestController {
-    private final LdesConfigModelService service;
+	private final LdesConfigModelService service;
 
 	@Autowired
 	@Qualifier("shapeShaclValidator")
@@ -30,17 +29,17 @@ public class AdminShapeRestController {
 		binder.setValidator(shapeValidator);
 	}
 
-    @GetMapping("/eventstreams/{collectionName}/shape")
-    public ResponseEntity<LdesConfigModel> getShape(@PathVariable String collectionName) {
-        LdesConfigModel shape = service.retrieveShape(collectionName);
-        return ResponseEntity.ok(shape);
-    }
+	@GetMapping("/eventstreams/{collectionName}/shape")
+	public ResponseEntity<LdesConfigModel> getShape(@PathVariable String collectionName) {
+		LdesConfigModel shape = service.retrieveShape(collectionName);
+		return ResponseEntity.ok(shape);
+	}
 
-    @PutMapping("/eventstreams/{collectionName}/shape")
-    public ResponseEntity<LdesConfigModel> putShape(@PathVariable String collectionName,
-                                                    @RequestBody @Validated LdesConfigModel shape) {
-        LdesConfigModel updatedShape = service.updateShape(collectionName, shape);
-        return ResponseEntity.ok(updatedShape);
-    }
+	@PutMapping("/eventstreams/{collectionName}/shape")
+	public ResponseEntity<LdesConfigModel> putShape(@PathVariable String collectionName,
+			@RequestBody @Validated LdesConfigModel shape) {
+		LdesConfigModel updatedShape = service.updateShape(collectionName, shape);
+		return ResponseEntity.ok(updatedShape);
+	}
 
 }
