+++ ldes-server-port-admin-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/exceptions/InvalidModelIdException.java
@@ -2,10 +2,10 @@ package be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.exceptions;
 
 import static be.vlaanderen.informatievlaanderen.ldes.server.domain.constants.RdfConstants.RDF_SYNTAX_TYPE;
 
-public class InvalidModelException extends RuntimeException {
+public class InvalidModelIdException extends RuntimeException {
 	private final String model;
 
-	public InvalidModelException(String model) {
+	public InvalidModelIdException(String model) {
 		this.model = model;
 	}
