+++ ldes-server-port-admin-rest/src/test/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/controllers/AdminShapeRestControllerTest.java
@@ -4,8 +4,8 @@ import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.config.AdminWeb
 import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.exceptionhandling.AdminRestResponseEntityExceptionHandler;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.converter.RdfModelConverter;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.exceptions.MissingLdesConfigException;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.services.LdesConfigModelService;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.valueobjects.LdesConfigModel;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.services.LdesConfigModelService;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.valueobjects.LdesConfigModel;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.validation.LdesConfigShaclValidator;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.riot.Lang;
@@ -32,7 +32,6 @@ import java.nio.file.Paths;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
@@ -44,98 +43,100 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @WebMvcTest
 @ActiveProfiles({ "test", "rest" })
 @ContextConfiguration(classes = { AdminShapeRestController.class,
-        AdminWebConfig.class, AdminRestResponseEntityExceptionHandler.class })
+		AdminWebConfig.class, AdminRestResponseEntityExceptionHandler.class })
 class AdminShapeRestControllerTest {
-    @MockBean
-    private LdesConfigModelService ldesConfigModelService;
-    @MockBean
-    @Qualifier("shapeShaclValidator")
-    private LdesConfigShaclValidator ldesConfigShaclValidator;
+	@MockBean
+	private LdesConfigModelService ldesConfigModelService;
+	@MockBean
+	@Qualifier("shapeShaclValidator")
+	private LdesConfigShaclValidator ldesConfigShaclValidator;
 
-    @Autowired
-    private MockMvc mockMvc;
+	@Autowired
+	private MockMvc mockMvc;
 
-    @BeforeEach
+	@BeforeEach
     void setUp() {
         when(ldesConfigShaclValidator.supports(any())).thenReturn(true);
     }
 
-    @Test
-    void when_StreamAndShapeArePresent_Then_ShapeIsReturned() throws Exception {
-        String collectionName = "name1";
-        Model expectedShapeModel = readModelFromFile("shape-1.ttl");
-        LdesConfigModel shapeConfigModel = new LdesConfigModel(collectionName + "Shape", expectedShapeModel);
-        when(ldesConfigModelService.retrieveShape(collectionName)).thenReturn(shapeConfigModel);
-        ResultActions resultActions = mockMvc.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/shape"))
-                .andDo(print())
-                .andExpect(status().isOk());
-        MvcResult result = resultActions.andReturn();
-        Model actualModel = RdfModelConverter.fromString(result.getResponse().getContentAsString(), Lang.TURTLE);
-        Assertions.assertTrue(actualModel.isIsomorphicWith(expectedShapeModel));
-    }
+	@Test
+	void when_StreamAndShapeArePresent_Then_ShapeIsReturned() throws Exception {
+		String collectionName = "name1";
+		Model expectedShapeModel = readModelFromFile("shape-1.ttl");
+		LdesConfigModel shapeConfigModel = new LdesConfigModel(collectionName + "Shape", expectedShapeModel);
+		when(ldesConfigModelService.retrieveShape(collectionName)).thenReturn(shapeConfigModel);
+		ResultActions resultActions = mockMvc.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/shape"))
+				.andDo(print())
+				.andExpect(status().isOk());
+		MvcResult result = resultActions.andReturn();
+		Model actualModel = RdfModelConverter.fromString(result.getResponse().getContentAsString(), Lang.TURTLE);
+		Assertions.assertTrue(actualModel.isIsomorphicWith(expectedShapeModel));
+	}
 
-    @Test
-    void when_ViewNotPresent_Then_Returned404() throws Exception {
-        String collectionName = "name1";
-        String viewName = "view1";
-        Model expectedViewModel = readModelFromFile("view-1.ttl");
-        LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
-        when(ldesConfigModelService.retrieveView(collectionName,viewName)).thenThrow(new MissingLdesConfigException(collectionName + "/" + viewName));
-        ResultActions resultActions = mockMvc.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
-                .andDo(print())
-                .andExpect(status().isNotFound());
-    }
+	@Test
+	void when_ViewNotPresent_Then_Returned404() throws Exception {
+		String collectionName = "name1";
+		String viewName = "view1";
+		Model expectedViewModel = readModelFromFile("view-1.ttl");
+		LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
+		when(ldesConfigModelService.retrieveView(collectionName, viewName))
+				.thenThrow(new MissingLdesConfigException(collectionName + "/" + viewName));
+		ResultActions resultActions = mockMvc
+				.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
+				.andDo(print())
+				.andExpect(status().isNotFound());
+	}
 
-    @Test
-    void when_ModelInRequestBody_Then_MethodIsCalled() throws Exception {
-        String collectionName = "name1";
-        String shapeName = "shape";
-        String fileName = "shape-1.ttl";
-        Model expectedShapeModel = readModelFromFile(fileName);
-        LdesConfigModel configModel = new LdesConfigModel(shapeName, expectedShapeModel);
-        when(ldesConfigModelService.addView(anyString(),any())).thenReturn(configModel);
-        mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
-                        .content(readDataFromFile(fileName, Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print())
-                .andExpect(status().isOk());
-        verify(ldesConfigModelService, times(1)).updateShape(anyString(), any());
-    }
+	@Test
+	void when_ModelInRequestBody_Then_MethodIsCalled() throws Exception {
+		String collectionName = "name1";
+		String shapeName = "shape";
+		String fileName = "shape-1.ttl";
+		Model expectedShapeModel = readModelFromFile(fileName);
+		LdesConfigModel configModel = new LdesConfigModel(shapeName, expectedShapeModel);
+		when(ldesConfigModelService.addView(anyString(), any())).thenReturn(configModel);
+		mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
+				.content(readDataFromFile(fileName, Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print())
+				.andExpect(status().isOk());
+		verify(ldesConfigModelService, times(1)).updateShape(anyString(), any());
+	}
 
-    @Test
-    void when_ModelWithoutType_Then_ReturnedBadRequest() throws Exception {
-        String collectionName = "name1";
-        ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
-                        .content(readDataFromFile("shape-without-type.ttl", Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-    }
+	@Test
+	void when_ModelWithoutType_Then_ReturnedBadRequest() throws Exception {
+		String collectionName = "name1";
+		ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
+				.content(readDataFromFile("shape-without-type.ttl", Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print())
+				.andExpect(status().isBadRequest());
+	}
 
-    @Test
-    void when_StreamEndpointCalledAndModelInRequestBody_Then_ModelIsValidated() throws Exception {
-        String collectionName = "name1";
-        String fileName = "shape-1.ttl";
-        mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
-                        .content(readDataFromFile(fileName, Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print());
-        verify(ldesConfigShaclValidator, times(1)).validate(any(), any());
-    }
+	@Test
+	void when_StreamEndpointCalledAndModelInRequestBody_Then_ModelIsValidated() throws Exception {
+		String collectionName = "name1";
+		String fileName = "shape-1.ttl";
+		mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/shape")
+				.content(readDataFromFile(fileName, Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print());
+		verify(ldesConfigShaclValidator, times(1)).validate(any(), any());
+	}
 
-    private String readDataFromFile(String fileName, Lang rdfFormat)
-            throws URISyntaxException, IOException {
-        ClassLoader classLoader = getClass().getClassLoader();
-        File file = new File(Objects.requireNonNull(classLoader.getResource(fileName)).toURI());
-        String content = Files.lines(Paths.get(file.toURI())).collect(Collectors.joining("\n"));
-        return content;
-    }
+	private String readDataFromFile(String fileName, Lang rdfFormat)
+			throws URISyntaxException, IOException {
+		ClassLoader classLoader = getClass().getClassLoader();
+		File file = new File(Objects.requireNonNull(classLoader.getResource(fileName)).toURI());
+		String content = Files.lines(Paths.get(file.toURI())).collect(Collectors.joining("\n"));
+		return content;
+	}
 
-    private Model readModelFromFile(String fileName) throws URISyntaxException {
-        ClassLoader classLoader = getClass().getClassLoader();
-        String uri = Objects.requireNonNull(classLoader.getResource(fileName)).toURI()
-                .toString();
-        return RDFDataMgr.loadModel(uri);
-    }
+	private Model readModelFromFile(String fileName) throws URISyntaxException {
+		ClassLoader classLoader = getClass().getClassLoader();
+		String uri = Objects.requireNonNull(classLoader.getResource(fileName)).toURI()
+				.toString();
+		return RDFDataMgr.loadModel(uri);
+	}
 
 }
