+++ ldes-server-port-admin-rest/src/test/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/controllers/AdminViewsRestControllerTest.java
@@ -1,12 +1,11 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.controllers;
 
 import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.config.AdminWebConfig;
-import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.converters.LdesConfigModelConverter;
 import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.exceptionhandling.AdminRestResponseEntityExceptionHandler;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.converter.RdfModelConverter;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.exceptions.MissingLdesConfigException;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.services.LdesConfigModelService;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.valueobjects.LdesConfigModel;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.services.LdesConfigModelService;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.ldesconfig.valueobjects.LdesConfigModel;
 import be.vlaanderen.informatievlaanderen.ldes.server.domain.validation.LdesConfigShaclValidator;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.riot.Lang;
@@ -33,7 +32,6 @@ import java.nio.file.Paths;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 import static org.mockito.Mockito.when;
@@ -45,102 +43,105 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @WebMvcTest
 @ActiveProfiles({ "test", "rest" })
 @ContextConfiguration(classes = { AdminViewsRestController.class,
-        AdminWebConfig.class, AdminRestResponseEntityExceptionHandler.class})
+		AdminWebConfig.class, AdminRestResponseEntityExceptionHandler.class })
 class AdminViewsRestControllerTest {
 
-    @MockBean
-    private LdesConfigModelService ldesConfigModelService;
-    @MockBean
-    @Qualifier("viewShaclValidator")
-    private LdesConfigShaclValidator ldesConfigShaclValidator;
+	@MockBean
+	private LdesConfigModelService ldesConfigModelService;
+	@MockBean
+	@Qualifier("viewShaclValidator")
+	private LdesConfigShaclValidator ldesConfigShaclValidator;
 
-    @Autowired
-    private MockMvc mockMvc;
+	@Autowired
+	private MockMvc mockMvc;
 
-    @BeforeEach
+	@BeforeEach
     void setUp() {
         when(ldesConfigShaclValidator.supports(any())).thenReturn(true);
     }
 
-    @Test
-    void when_StreamAndViewArePresent_Then_ViewIsReturned() throws Exception {
-        String collectionName = "name1";
-        String viewName = "view1";
-        Model expectedViewModel = readModelFromFile("view-1.ttl");
-        LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
-        when(ldesConfigModelService.retrieveView(collectionName,viewName)).thenReturn(configModel);
-        ResultActions resultActions = mockMvc.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
-                .andDo(print())
-                .andExpect(status().isOk());
-        MvcResult result = resultActions.andReturn();
-        Model actualModel = RdfModelConverter.fromString(result.getResponse().getContentAsString(), Lang.TURTLE);
-        Assertions.assertTrue(actualModel.isIsomorphicWith(expectedViewModel));
-    }
+	@Test
+	void when_StreamAndViewArePresent_Then_ViewIsReturned() throws Exception {
+		String collectionName = "name1";
+		String viewName = "view1";
+		Model expectedViewModel = readModelFromFile("view-1.ttl");
+		LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
+		when(ldesConfigModelService.retrieveView(collectionName, viewName)).thenReturn(configModel);
+		ResultActions resultActions = mockMvc
+				.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
+				.andDo(print())
+				.andExpect(status().isOk());
+		MvcResult result = resultActions.andReturn();
+		Model actualModel = RdfModelConverter.fromString(result.getResponse().getContentAsString(), Lang.TURTLE);
+		Assertions.assertTrue(actualModel.isIsomorphicWith(expectedViewModel));
+	}
 
-    @Test
-    void when_ViewNotPresent_Then_Returned404() throws Exception {
-        String collectionName = "name1";
-        String viewName = "view1";
-        Model expectedViewModel = readModelFromFile("view-1.ttl");
-        LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
-        when(ldesConfigModelService.retrieveView(collectionName,viewName)).thenThrow(new MissingLdesConfigException(collectionName + "/" + viewName));
-        ResultActions resultActions = mockMvc.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
-                .andDo(print())
-                .andExpect(status().isNotFound());
-    }
+	@Test
+	void when_ViewNotPresent_Then_Returned404() throws Exception {
+		String collectionName = "name1";
+		String viewName = "view1";
+		Model expectedViewModel = readModelFromFile("view-1.ttl");
+		LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
+		when(ldesConfigModelService.retrieveView(collectionName, viewName))
+				.thenThrow(new MissingLdesConfigException(collectionName + "/" + viewName));
+		ResultActions resultActions = mockMvc
+				.perform(get("/admin/api/v1/eventstreams/" + collectionName + "/views/" + viewName))
+				.andDo(print())
+				.andExpect(status().isNotFound());
+	}
 
-    @Test
-    void when_ModelInRequestBody_Then_MethodIsCalled() throws Exception {
-        String collectionName = "name1";
-        String viewName = "view1";
-        Model expectedViewModel = readModelFromFile("view-1.ttl");
-        LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
-        when(ldesConfigModelService.addView(anyString(),any())).thenReturn(configModel);
-        ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
-                        .content(readDataFromFile("view-1.ttl", Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print())
-                .andExpect(status().isOk());
-        verify(ldesConfigModelService, times(1)).addView(anyString(), any());
-    }
+	@Test
+	void when_ModelInRequestBody_Then_MethodIsCalled() throws Exception {
+		String collectionName = "name1";
+		String viewName = "view1";
+		Model expectedViewModel = readModelFromFile("view-1.ttl");
+		LdesConfigModel configModel = new LdesConfigModel(viewName, expectedViewModel);
+		when(ldesConfigModelService.addView(anyString(), any())).thenReturn(configModel);
+		ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
+				.content(readDataFromFile("view-1.ttl", Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print())
+				.andExpect(status().isOk());
+		verify(ldesConfigModelService, times(1)).addView(anyString(), any());
+	}
 
-    @Test
-    void when_ModelWithoutType_Then_ReturnedBadRequest() throws Exception {
-        String collectionName = "name1";
-        ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
-                        .content(readDataFromFile("view-without-type.ttl", Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print())
-                .andExpect(status().isBadRequest());
-    }
+	@Test
+	void when_ModelWithoutType_Then_ReturnedBadRequest() throws Exception {
+		String collectionName = "name1";
+		ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
+				.content(readDataFromFile("view-without-type.ttl", Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print())
+				.andExpect(status().isBadRequest());
+	}
 
-    @Test
-    void when_StreamEndpointCalledAndModelInRequestBody_Then_ModelIsValidated() throws Exception {
-        String collectionName = "name1";
-        String viewName = "view1";
-        final Model model = readModelFromFile("view-1.ttl");
-        final LdesConfigModel ldesConfigModel = new LdesConfigModel(viewName, model);
-        when(ldesConfigModelService.addView(collectionName, ldesConfigModel)).thenReturn(ldesConfigModel);
-        ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
-                        .content(readDataFromFile("ldes-1.ttl", Lang.TURTLE))
-                        .contentType(MediaType.TEXT_PLAIN))
-                .andDo(print());
-        verify(ldesConfigShaclValidator, times(1)).validate(any(), any());
-    }
+	@Test
+	void when_StreamEndpointCalledAndModelInRequestBody_Then_ModelIsValidated() throws Exception {
+		String collectionName = "name1";
+		String viewName = "view1";
+		final Model model = readModelFromFile("view-1.ttl");
+		final LdesConfigModel ldesConfigModel = new LdesConfigModel(viewName, model);
+		when(ldesConfigModelService.addView(collectionName, ldesConfigModel)).thenReturn(ldesConfigModel);
+		ResultActions resultActions = mockMvc.perform(put("/admin/api/v1/eventstreams/" + collectionName + "/views")
+				.content(readDataFromFile("ldes-1.ttl", Lang.TURTLE))
+				.contentType(MediaType.TEXT_PLAIN))
+				.andDo(print());
+		verify(ldesConfigShaclValidator, times(1)).validate(any(), any());
+	}
 
-    private String readDataFromFile(String fileName, Lang rdfFormat)
-            throws URISyntaxException, IOException {
-        ClassLoader classLoader = getClass().getClassLoader();
-        File file = new File(Objects.requireNonNull(classLoader.getResource(fileName)).toURI());
-        String content = Files.lines(Paths.get(file.toURI())).collect(Collectors.joining("\n"));
-        return content;
-    }
+	private String readDataFromFile(String fileName, Lang rdfFormat)
+			throws URISyntaxException, IOException {
+		ClassLoader classLoader = getClass().getClassLoader();
+		File file = new File(Objects.requireNonNull(classLoader.getResource(fileName)).toURI());
+		String content = Files.lines(Paths.get(file.toURI())).collect(Collectors.joining("\n"));
+		return content;
+	}
 
-    private Model readModelFromFile(String fileName) throws URISyntaxException {
-        ClassLoader classLoader = getClass().getClassLoader();
-        String uri = Objects.requireNonNull(classLoader.getResource(fileName)).toURI()
-                .toString();
-        return RDFDataMgr.loadModel(uri);
-    }
+	private Model readModelFromFile(String fileName) throws URISyntaxException {
+		ClassLoader classLoader = getClass().getClassLoader();
+		String uri = Objects.requireNonNull(classLoader.getResource(fileName)).toURI()
+				.toString();
+		return RDFDataMgr.loadModel(uri);
+	}
 
 }
